--- C code/Resistor Color Code/.github/workflows/CodeQuality.yml
@@ -1,22 +1,0 @@
-name: CodeQuality Dynamic Code Analysis Valgrind
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  test:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: apt install dependency
-      run: |
-        sudo apt-get -y install valgrind
-        sudo apt-get -y install libcunit1 libcunit1-doc libcunit1-dev 
-    - name: make test
-      run: make -C 3_Implementation/ test
-    - name: Valgrid
-      run: valgrind ./3_Implementation/build/Code.out.out
