--- C code/Resistor Color Code/.github/workflows/cppcheck.yml
@@ -1,19 +1,0 @@
-name: Code Quality - Static Code - Cppcheck
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  cppcheck:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Install cppcheck
-      run: sudo apt -y install cppcheck
-    - name: Run cppcheck
-      run: cppcheck 3_Implementation
