--- C code/Resistor Color Code/.github/workflows/gitinspector.yml
@@ -1,51 +1,0 @@
-name: Contribution Check - Git Inspector
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  inspect:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-      with:
-        fetch-depth: '0'
-    - name: Install python
-      run: sudo apt -y install python
-    - name: Install gitinspector
-      run: | 
-        pushd ~ &> /dev/null
-        git clone --depth=1 https://github.com/ejwa/gitinspector.git
-        cd gitinspector
-        echo "export PATH=$(pwd):"'$PATH' >> ~/.profile
-        popd &> /dev/null
-    - name: Run gitinspector
-      run: |
-        source ~/.profile
-        export PYTHONIOENCODING=utf-8
-        gitinspector.py --version
-        echo "Running for commit : $GITHUB_SHA"
-        echo "============ Git Inspector ==========="
-        gitinspector.py -wTHl --format=text -f c,h,md,yml,* -x file:3_implementation/unity -x file:doc .
-        gitinspector.py -wTHl --format=html -f c,h,md,yml,* -x file:3_implementation/unity -x file:doc . > ~/result.html
-        echo "======================================"
-   #If using gihub io pages feature
-   # - name : Update Results to gh-pages
-   #   uses: actions/checkout@v2
-   #   with:
-   #     ref: gh-pages
-   # - name: Update Results to gh-pages
-   #  run: |
-   #     pwd
-   #     rm -f result.html
-   #     mv ~/result.html result.html
-   #     git config user.name "GitHub Actions"
-   #     git config user.email actions@github.com
-   #     git add result.html
-   #     git commit -m "[auto] Update Results for commit SHA $GITHUB_SHA"
-   #     git push
