--- C code/Resistor Color Code/.github/workflows/main.yml
@@ -1,17 +1,0 @@
-name: Unit Testing - Unity
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  test:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: make
-      run: make -C 3_Implementation/ test
