--- C code/Resistor Color Code/2_Design/Readme.md
@@ -1,20 +1,0 @@
-# Design
-
-## High Level Design 
-
-### Structural Diagram
-<img src="https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/blob/master/2_Design/hight%20level%20struct.png" >
-
-### Behavioral Diagram
-<img src="https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/blob/master/2_Design/High%20Level%20Behave.png" > 
-
-
-## Low Level Design 
-
-### Structural Diagram
-<img src="https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/blob/master/2_Design/Low%20level%20struct.png" >
-
-
-### Behavioral Diagram 
-<img src="https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/blob/master/2_Design/Lowlevel%20Behave.png" >
-
