--- C code/Resistor Color Code/3_Implementation/Readme.md
@@ -1,10 +1,0 @@
-# Implementation
-
-## Folder Structure
-Folder        | description
---------------| ----------------------------------------------
-`build`       | Build output (Not included in git)
-`inc`         | All header files
-`src`         | Main source code for calculator
-`test`        | All source code and data for testing purposes
-`unity`       | All header files for testing purposes
