--- C code/Resistor Color Code/3_Implementation/inc/header.h
@@ -1,12 +1,0 @@
-#ifndef _HEADER_H_
-#define _HEADER_H_
-
-char *DigitSeparate (char *ret, int inp); //seprates unit digits
-
-int Color (char *value);                  //search the index values in color codes
-
-char *Converter (char *trgt, char *src);   //converts upper case elements to lower case
-
-int ColorMatch (char *color1, char *color2); //comparison between two color
-
-#endif
