--- C code/Resistor Color Code/4_TestPlan/readme.md
@@ -1,15 +1,0 @@
-# TEST PLAN:
-
-## Table no: High level test plan
-
-| **Test ID** | **Description**                                              | **Exp I/P** | **Exp O/P** | **Actual Out** |**Type Of Test**  |    
-|-------------|--------------------------------------------------------------|------------|-------------|----------------|------------------|
-|  H_01    | Check if the passed colouris valid|Valid Resistance |Matched |Matched|Passed |
-|  H_02       |Check if the passed resistance colour is returning the value|Valid Resistance Colour code input|Returned and Matched|Matched|Passed|
-
-## Table no: Low level test plan
-
-| **Test ID** | **Description**                                              | **Exp IN** | **Exp OUT** | **Actual Out** |**Type Of Test**  |    
-|-------------|--------------------------------------------------------------|------------|-------------|----------------|------------------|
-|  L_01       |Check if the passed Resistance shaded code is substantial to the matchcolour work|Valid Resistance Colour|  Valid |Matched|PASSED|
-|  L_02       |Check if the passed shaded code is legitimate to Search_BandColour work and returning its individual Color code| Valid Resistance|Matched|Returned Code |PASSED|
