--- C code/Resistor Color Code/README.md
@@ -1,20 +1,0 @@
-# Resistor Color Decoding
-
-[![Code Quality - Static Code - Cppcheck](https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/actions/workflows/cppcheck.yml/badge.svg)](https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/actions/workflows/cppcheck.yml)
-[![Code Quality - Static Code - Cppcheck](https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/actions/workflows/cppcheck.yml/badge.svg)](https://github.com/VisheshYadav288037/Advanced-Electrical-Electronics/actions/workflows/cppcheck.yml)
-
-<div align="center">
-<img src="https://eepower.com/uploads/education/resistor_color_codes_chart.png" >
-</div>
-
-## Folder Structure
-Folder             | Description
--------------------| -----------------------------------------
-`1_Requirements`   | Documents detailing requirements and research
-`2_Design`         | Documents specifying design details
-`3_Implementation` | All code and documentation
-`4_Test_plan`      | Documents with test plans and procedures
-
-
-
-
